requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0

# Data handling
pyyaml>=6.0
pillow>=9.0.0
numpy==1.24.3

# Metrics and evaluation
scikit-learn>=1.3.0

# Development and testing
flake8>=6.0.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
isort>=5.12.0